eloapp-13 wydzielenie RatingsHistorySvc
